requirements.txt
django
psycopg2
django-seed
djangorestframework
django-filter
drf-yasg